+++ wayland-scanner/src/client_gen.rs
@@ -246,7 +246,10 @@ fn gen_methods(interface: &Interface) -> TokenStream {
                 quote! {
                     #doc_attr
                     #[allow(clippy::too_many_arguments)]
-                    pub fn #method_name<U: Send + Sync + 'static, D: Dispatch<super::#created_iface_mod::#created_iface_type, U> + 'static>(&self, #(#fn_args,)* qh: &QueueHandle<D>, udata: U) -> super::#created_iface_mod::#created_iface_type {
+                    pub fn #method_name<U, D: 'static>(&self, #(#fn_args,)* qh: &QueueHandle<D>, udata: U) -> super::#created_iface_mod::#created_iface_type
+                    where
+                        U: Dispatch<super::#created_iface_mod::#created_iface_type, D> + Send + Sync + 'static
+                    {
                         self.send_constructor(
                             Request::#enum_variant {
                                 #(#enum_args),*
@@ -261,7 +264,10 @@ fn gen_methods(interface: &Interface) -> TokenStream {
                 quote! {
                     #doc_attr
                     #[allow(clippy::too_many_arguments)]
-                    pub fn #method_name<I: Proxy + 'static, U: Send + Sync + 'static, D: Dispatch<I, U> + 'static>(&self, #(#fn_args,)* qh: &QueueHandle<D>, udata: U) -> I {
+                    pub fn #method_name<I: Proxy + 'static, U, D: 'static>(&self, #(#fn_args,)* qh: &QueueHandle<D>, udata: U) -> I
+                    where
+                        U: Dispatch<I, D> + Send + Sync + 'static
+                    {
                         self.send_constructor(
                             Request::#enum_variant {
                                 #(#enum_args),*