+++ wayland-tests/tests/client_bad_requests.rs
@@ -12,7 +12,8 @@ fn constructor_dead() {
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: globals::GlobalList::new() };
 
-    let registry = client.display.get_registry(&client.event_queue.handle(), ());
+    let registry =
+        client.display.get_registry(&client.event_queue.handle(), globals::GlobalListData);
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
 
@@ -22,13 +23,13 @@ fn constructor_dead() {
             &client.event_queue.handle(),
             &registry,
             1..=1,
-            (),
+            wayc::NoopIgnore,
         )
         .unwrap();
 
     seat.release();
 
-    assert!(seat.get_pointer(&client.event_queue.handle(), ()).id().is_null());
+    assert!(seat.get_pointer(&client.event_queue.handle(), wayc::NoopIgnore).id().is_null());
 }
 
 #[test]
@@ -42,7 +43,8 @@ fn send_constructor_wrong_type() {
     let (_, mut client) = server.add_client();
     let mut client_ddata = ClientHandler { globals: globals::GlobalList::new() };
 
-    let registry = client.display.get_registry(&client.event_queue.handle(), ());
+    let registry =
+        client.display.get_registry(&client.event_queue.handle(), globals::GlobalListData);
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut ServerHandler).unwrap();
 
@@ -52,7 +54,7 @@ fn send_constructor_wrong_type() {
             &client.event_queue.handle(),
             &registry,
             1..=1,
-            (),
+            wayc::NoopIgnore,
         )
         .unwrap();
 
@@ -65,7 +67,7 @@ fn send_constructor_wrong_type() {
                 client
                     .event_queue
                     .handle()
-                    .make_data::<wayc::protocol::wl_keyboard::WlKeyboard, _>(()),
+                    .make_data::<wayc::protocol::wl_keyboard::WlKeyboard, _>(wayc::NoopIgnore),
             ),
         )
         .unwrap();
@@ -87,16 +89,6 @@ impl AsMut<globals::GlobalList> for ClientHandler {
         &mut self.globals
     }
 }
-
-wayc::delegate_dispatch!(ClientHandler:
-    [wayc::protocol::wl_registry::WlRegistry: ()] => globals::GlobalList
-);
-
-client_ignore_impl!(ClientHandler => [
-    wayc::protocol::wl_seat::WlSeat,
-    wayc::protocol::wl_pointer::WlPointer,
-    wayc::protocol::wl_keyboard::WlKeyboard
-]);
 
 /*
  * Server handler