+++ wayland-tests/tests/server_global_filter.rs
@@ -1,6 +1,4 @@
-use wayland_tests::{
-    TestServer, client_ignore_impl, globals, roundtrip, server_ignore_impl, wayc, ways,
-};
+use wayland_tests::{TestServer, globals, roundtrip, server_ignore_impl, wayc, ways};
 
 use ways::protocol::{wl_compositor, wl_output, wl_shm};
 
@@ -25,7 +23,7 @@ fn global_filter() {
     let (_, mut client) = server.add_client_with_data(Arc::new(MyClientData { privileged: false }));
     let mut client_ddata = ClientHandler::new();
 
-    client.display.get_registry(&client.event_queue.handle(), ());
+    client.display.get_registry(&client.event_queue.handle(), globals::GlobalListData);
 
     roundtrip(&mut client, &mut server, &mut client_ddata, &mut server_ddata).unwrap();
 
@@ -35,7 +33,7 @@ fn global_filter() {
         server.add_client_with_data(Arc::new(MyClientData { privileged: true }));
     let mut priv_client_ddata = ClientHandler::new();
 
-    priv_client.display.get_registry(&priv_client.event_queue.handle(), ());
+    priv_client.display.get_registry(&priv_client.event_queue.handle(), globals::GlobalListData);
 
     roundtrip(&mut priv_client, &mut server, &mut priv_client_ddata, &mut server_ddata).unwrap();
 
@@ -76,22 +74,25 @@ fn global_filter_try_force() {
 
     // privileged client can bind it
 
-    let priv_registry = priv_client.display.get_registry(&priv_client.event_queue.handle(), ());
+    let priv_registry = priv_client
+        .display
+        .get_registry(&priv_client.event_queue.handle(), globals::GlobalListData);
     priv_registry.bind::<wayc::protocol::wl_output::WlOutput, _, _>(
         1,
         1,
         &priv_client.event_queue.handle(),
-        (),
+        wayc::NoopIgnore,
     );
     roundtrip(&mut priv_client, &mut server, &mut priv_client_ddata, &mut server_ddata).unwrap();
 
     // unprivileged client cannot
-    let registry = client.display.get_registry(&client.event_queue.handle(), ());
+    let registry =
+        client.display.get_registry(&client.event_queue.handle(), globals::GlobalListData);
     registry.bind::<wayc::protocol::wl_output::WlOutput, _, _>(
         1,
         1,
         &client.event_queue.handle(),
-        (),
+        wayc::NoopIgnore,
     );
 
     assert!(roundtrip(&mut client, &mut server, &mut client_ddata, &mut server_ddata).is_err());
@@ -113,65 +114,47 @@ impl AsMut<globals::GlobalList> for ClientHandler {
     }
 }
 
-wayc::delegate_dispatch!(ClientHandler:
-    [wayc::protocol::wl_registry::WlRegistry: ()] => globals::GlobalList
-);
-
-client_ignore_impl!(ClientHandler => [
-    wayc::protocol::wl_compositor::WlCompositor,
-    wayc::protocol::wl_shm::WlShm,
-    wayc::protocol::wl_output::WlOutput
-]);
-
 struct ServerHandler;
 
-impl ways::GlobalDispatch<wl_compositor::WlCompositor, ()> for ServerHandler {
+impl ways::GlobalDispatch<wl_compositor::WlCompositor, ServerHandler> for () {
     fn bind(
-        _: &mut Self,
+        &self,
+        _: &mut ServerHandler,
         _: &ways::DisplayHandle,
         _: &ways::Client,
         resource: ways::New<wl_compositor::WlCompositor>,
-        _: &(),
-        data_init: &mut ways::DataInit<'_, Self>,
+        data_init: &mut ways::DataInit<'_, ServerHandler>,
     ) {
         data_init.init(resource, ());
     }
-
-    fn can_view(_: ways::Client, _: &()) -> bool {
-        true
-    }
 }
 
-impl ways::GlobalDispatch<wl_shm::WlShm, ()> for ServerHandler {
+impl ways::GlobalDispatch<wl_shm::WlShm, ServerHandler> for () {
     fn bind(
-        _: &mut Self,
+        &self,
+        _: &mut ServerHandler,
         _: &ways::DisplayHandle,
         _: &ways::Client,
         resource: ways::New<wl_shm::WlShm>,
-        _: &(),
-        data_init: &mut ways::DataInit<'_, Self>,
+        data_init: &mut ways::DataInit<'_, ServerHandler>,
     ) {
         data_init.init(resource, ());
-    }
-
-    fn can_view(_: ways::Client, _: &()) -> bool {
-        true
     }
 }
 
-impl ways::GlobalDispatch<wl_output::WlOutput, ()> for ServerHandler {
+impl ways::GlobalDispatch<wl_output::WlOutput, ServerHandler> for () {
     fn bind(
-        _: &mut Self,
+        &self,
+        _: &mut ServerHandler,
         _: &ways::DisplayHandle,
         _: &ways::Client,
         resource: ways::New<wl_output::WlOutput>,
-        _: &(),
-        data_init: &mut ways::DataInit<'_, Self>,
+        data_init: &mut ways::DataInit<'_, ServerHandler>,
     ) {
         data_init.init(resource, ());
     }
 
-    fn can_view(client: ways::Client, _: &()) -> bool {
+    fn can_view(&self, client: &ways::Client) -> bool {
         client.get_data::<MyClientData>().unwrap().privileged
     }
 }